chore: organize common folder
