+++ src/common/decorators/roles.decorator.ts
@@ -1,4 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
+
 import { Role } from '../enums/role.enum';
 
 export const ROLES_KEY = 'roles';
