+++ src/common/dtos/auth/login.dto.ts
@@ -1,0 +1,14 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsString, IsNotEmpty } from 'class-validator';
+
+export class LoginDto {
+  @ApiProperty()
+  @IsString()
+  @IsNotEmpty()
+  readonly username: string;
+
+  @ApiProperty()
+  @IsString()
+  @IsNotEmpty()
+  readonly password: string;
+}
