+++ src/common/dtos/users/create-user.dto.ts
@@ -1,0 +1,14 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsString, IsNotEmpty } from 'class-validator';
+
+export class CreateUserDto {
+  @ApiProperty()
+  @IsString()
+  @IsNotEmpty()
+  readonly password: string;
+
+  @ApiProperty()
+  @IsString()
+  @IsNotEmpty()
+  readonly username: string;
+}
