+++ src/common/responses/auth/tokens.response.ts
@@ -1,19 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsString, IsNotEmpty } from 'class-validator';
-import { Role } from '../enums/role.enum';
 
-export class LoginDto {
-  @ApiProperty()
-  @IsString()
-  readonly username: string;
-
-  @ApiProperty()
-  @IsString()
-  @IsNotEmpty()
-  readonly password: string;
-}
-
-export class LoginResponse {
+export class TokensResponse {
   @ApiProperty({
     example:
       'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c',
@@ -25,9 +12,4 @@ export class LoginResponse {
       'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c',
   })
   readonly refreshToken: string;
-}
-
-export interface PayloadToken {
-  id: string;
-  role: Role;
 }
