+++ src/common/types/auth.type.ts
@@ -1,0 +1,6 @@
+import { Role } from '../enums/role.enum';
+
+export interface PayloadToken {
+  id: string;
+  role: Role;
+}
